fix tests autodiscovery
